fix `Uncaught [Error: Invalid hook call.` (#6965)
* fix test1:unit

* hoist react & react-dom

this way we avoid possibly duplicated react & react-dom packages in
nested `node_modules`.

If duplication would occur, react is likely to throw an error
it is explained in detail here https://reactjs.org/warnings/invalid-hook-call-warning.html

* add correcter type annotation

otherwise typescript complains with error:

```
The inferred type of 'tooltipTestkitFactory' cannot be named without a
reference to wix-ui-core/node_modules/wix-ui-test-utils/protractor'. This is likely not portable. A type annotation is necessary.
```

* add `react-autodocs-utils` devDep to wix-style-react

it is used already but was not explicitly added. It came with
`wix-storybook-utils` and was assumed that it sits in resolvable
node_modules folder

* add npm dedupe

* dedupe only wix-style-react package

* disable hoisting, add `storybook/addons` as devDep

depends on https://github.com/wix/wix-ui/pull/2068